requirements.txt
lxml^=3.7.3
python-decouple^=3.1
requests^=2.10.0
urllib3^=1.15.1
